// src/interfaces/admin/user.interface.ts
export interface IUser {
  id: number;
  firstName: string;
  middleName: string;
  lastName: string;
  contactNumber: string;
  whatsAppNumber: string;
  dateOfBirthEnglish: string;
  dateOfBirthMisri: string;
  streetAddress: string;
  city: string;
  state: string;
  country: string;
  zipCode: string;
  email: string;
  isEmailVerified: boolean;
  role: string;
  createdAt: string;
  updatedAt: string;
}

export interface IUserUpdate {
  firstName: string;
  middleName: string;
  lastName: string;
  contactNumber: string;
  whatsAppNumber: string;
  dateOfBirthEnglish: string;
  dateOfBirthMisri: string;
  streetAddress: string;
  city: string;
  state: string;
  country: string;
  zipCode: string;
}